Remove unnecessary modules and rename some other
