+++ diffsky/data_loaders/hacc_utils/tests/test_lc_mock.py
@@ -18,10 +18,10 @@ from diffsky.param_utils import diffsky_param_wrapper as dpw
 from ....experimental import precompute_ssp_phot as psspp
 from ....experimental.disk_bulge_modeling import disk_bulge_kernels as dbk
 from ....experimental.disk_bulge_modeling import mc_disk_bulge as mcdb
-from ....experimental.lc_phot_kern_repro import get_wave_eff_table
+from ....experimental.lc_phot_kern import get_wave_eff_table
 from ....experimental.tests import test_mc_lightcone_halos as tmclh
 from ... import io_utils as iou
-from .. import lc_mock_repro as lcmp_repro
+from .. import lc_mock as lcmp_repro
 from .. import load_lc_cf
 
 vmap_interp = jjit(vmap(jnp.interp, in_axes=(0, None, 0)))
