+++ diffsky/experimental/kernels/ssp_weight_kernels.py
