+++ diffsky/experimental/tests/test_mc_phot.py
@@ -11,7 +11,7 @@ from dsps.photometry import photometry_kernels as phk
 from jax import random as jran
 from jax import vmap
 
-from .. import mc_phot_repro
+from .. import mc_phot
 from . import test_mc_lightcone_halos as tmclh
 
 _A = [None, 0, None, None, 0, *[None] * 4]
@@ -21,10 +21,10 @@ calc_obs_mags_galpop = vmap(phk.calc_obs_mag, in_axes=_A)
 def test_mc_lc_phot_changes_with_diffstarpop(num_halos=50):
     ran_key = jran.key(0)
     lc_data, tcurves = tmclh._get_weighted_lc_data_for_unit_testing(num_halos=num_halos)
-    phot_kern_results = mc_phot_repro.mc_lc_phot(
+    phot_kern_results = mc_phot.mc_lc_phot(
         ran_key, lc_data, diffstarpop_params=sfh_models["tng"]
     )
-    phot_kern_results2 = mc_phot_repro.mc_lc_phot(
+    phot_kern_results2 = mc_phot.mc_lc_phot(
         ran_key, lc_data, diffstarpop_params=sfh_models["smdpl_dr1"]
     )
     assert not np.allclose(
@@ -41,8 +41,8 @@ def test_mc_lc_phot_changes_with_diffstarpop(num_halos=50):
 def test_mc_lc_sed_is_consistent_with_mc_lc_phot(num_halos=50):
     ran_key = jran.key(0)
     lc_data, tcurves = tmclh._get_weighted_lc_data_for_unit_testing(num_halos=num_halos)
-    phot_kern_results = mc_phot_repro.mc_lc_phot(ran_key, lc_data)
-    sed_kern_results = mc_phot_repro.mc_lc_sed(ran_key, lc_data)
+    phot_kern_results = mc_phot.mc_lc_phot(ran_key, lc_data)
+    sed_kern_results = mc_phot.mc_lc_sed(ran_key, lc_data)
 
     phot_kern_results = namedtuple("Results", list(phot_kern_results.keys()))(
         **phot_kern_results
@@ -73,7 +73,7 @@ def test_mc_lc_sed_is_consistent_with_mc_lc_phot(num_halos=50):
 def test_mc_lc_dbk_phot(num_halos=50):
     ran_key = jran.key(0)
     lc_data, tcurves = tmclh._get_weighted_lc_data_for_unit_testing(num_halos=num_halos)
-    dbk_phot_info = mc_phot_repro.mc_lc_dbk_phot(ran_key, lc_data)
+    dbk_phot_info = mc_phot.mc_lc_dbk_phot(ran_key, lc_data)
 
     np.all(dbk_phot_info["logsm_obs"] > np.log10(dbk_phot_info["mstar_bulge"]))
     np.all(dbk_phot_info["logsm_obs"] > np.log10(dbk_phot_info["mstar_disk"]))
@@ -123,7 +123,7 @@ def check_phot_kern_results(phot_kern_results):
 def test_mc_lc_dbk_sed(num_halos=50):
     ran_key = jran.key(0)
     lc_data, tcurves = tmclh._get_weighted_lc_data_for_unit_testing(num_halos=num_halos)
-    dbk_sed_info = mc_phot_repro.mc_lc_dbk_sed(ran_key, lc_data)
+    dbk_sed_info = mc_phot.mc_lc_dbk_sed(ran_key, lc_data)
     assert np.all(np.isfinite(dbk_sed_info["rest_sed_bulge"]))
     assert np.all(np.isfinite(dbk_sed_info["rest_sed_disk"]))
     assert np.all(np.isfinite(dbk_sed_info["rest_sed_knots"]))
